monitoring-map/requirements.txt
numpy
pandas
matplotlib
shapely
scipy
pymongo
python-dotenv